+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+
+## 1.2.0 (2019-09-17)
+This release contains backwards incompatible changes.
+
+Changes on `OrderLineDisplayBasicSerializer`   
+  * Removed: `geo_targeting` field.
+  * New: `geo_targeting_zip` field which accepts one string zip code. Not mandatory
+  * New: `geo_targeting_radius` field accepting integer values between 1-80 (km). Mandatory only if `geo_targeting_zip` is given.  
+  * Removed: `stock_images_allowed` field:
+  * New: `banner_images_source` choice field accepting following values: 
+    * 0 for 'From website' / 'Von der Webseite' option
+    * 1 for 'From customer' / 'Vom kunden' option
+    * 2 for 'Customer photos' / 'Regiohelden Bilder' option 
+  * Change: `target_page_type` existing field which is not required anymore.
+
+Changes on `OrderLineGoogleAdsPremiumSerializer`:
+  * New: `call_tracking` boolean required field.
+  
+
 ## 1.1.3 (2019-09-11)
 Clean README file.